refactor(gateway): isolate internal agent facade contracts
